Report status from utils in a library
