+++ actions/store-logout
@@ -1,6 +1,10 @@
 #!/usr/bin/env python3
 
 import subprocess
+import sys
+sys.path.append('lib')
+from utils import report_status
+
 
 USER = "jenkins"
 
@@ -12,8 +16,7 @@ def charmstore_logout():
     cmd = ['su', USER, '-c', 'charm logout']
     if subprocess.call(cmd) != 0:
         raise Exception
-
-    subprocess.call(["hooks/update-status"])
+    report_status()
 
 
 if __name__ == "__main__":
