+++ actions/unregister-controller
@@ -11,6 +11,7 @@ from cwrhelpers import wait_result  # noqa: E402
 from charmhelpers.core import hookenv  # noqa: E402
 from charms.reactive import RelationBase  # noqa: E402
 from jenkins import Jenkins  # noqa: E402
+from utils import report_status
 
 
 def unregister_controller():
@@ -34,7 +35,7 @@ def unregister_controller():
     build_outcome, build_output = wait_result(jclient, "UnregisterController", next_build_number)
     hookenv.action_set({'outcome': build_outcome})
     hookenv.action_set({"output": build_output})
-    subprocess.call(["hooks/update-status"])
+    report_status()
 
 
 if __name__ == "__main__":
