+++ reactive/cwr.py
@@ -2,6 +2,7 @@ import os
 from pathlib import Path
 from tempfile import TemporaryDirectory
 import time
+from utils import report_status
 from charmhelpers import fetch
 from charmhelpers.core import host, hookenv
 from charms.reactive import (
@@ -17,35 +18,6 @@ from jujubigdata import utils
 from jenkins import Jenkins
 from CIGateway import CIGateway
 from controller import helpers
-
-
-def report_status():
-    if not is_state('jenkins.available'):
-        hookenv.status_set('waiting',
-                           'Waiting for jenkins to become available.')
-        return
-
-    # jenkins.available is set from here on
-    if not is_state('jenkins.jobs.ready'):
-        hookenv.status_set('waiting',
-                           'Waiting for jenkins jobs to be uploaded.')
-        return
-
-    # jenkins.available and jenkins.jobs.ready are set from here on
-    controllers = helpers.get_controllers()
-    if len(controllers) == 0:
-        hookenv.status_set('blocked',
-                           'Waiting for controller registration.')
-        return
-
-    # jenkins.available and jenkins.jobs.ready and controllers > 0 from here on
-    if helpers.get_charmstore_token():
-        msg = ('Ready (controllers: {}; store: authenticated).'
-               .format(controllers))
-    else:
-        msg = ('Ready (controllers: {}; store: unauthenticated).'
-               .format(controllers))
-    hookenv.status_set('active', msg)
 
 
 def pip_install_from_git(version, wheelhouse, repo):
